producer/requirements.txt
mysql-connector-python
selenium
webdriver_manager
python-dotenv